test(auto-reply): accelerate fallback coverage
