+++ src/auto-reply/reply/agent-runner-auth-profile.ts
@@ -41,3 +41,33 @@ export function resolveRunAuthProfile(
     workspaceDir: run.workspaceDir,
   });
 }
+
+/** Applies an auto-fallback probe's pinned auth to its fallback candidate. */
+export function resolveFallbackCandidateRun(
+  run: FollowupRun["run"],
+  provider: string,
+  model: string,
+): FollowupRun["run"] {
+  const probe = run.autoFallbackPrimaryProbe;
+  const isPrimaryProbeCandidate = probe && provider === probe.provider && model === probe.model;
+  if (
+    !probe ||
+    provider !== probe.fallbackProvider ||
+    isPrimaryProbeCandidate ||
+    !probe.fallbackAuthProfileId
+  ) {
+    return run;
+  }
+  const candidateRun: FollowupRun["run"] = {
+    ...run,
+    provider,
+    model,
+    authProfileId: probe.fallbackAuthProfileId,
+  };
+  if (probe.fallbackAuthProfileIdSource) {
+    candidateRun.authProfileIdSource = probe.fallbackAuthProfileIdSource;
+  } else {
+    delete candidateRun.authProfileIdSource;
+  }
+  return candidateRun;
+}
